Refactor API directory by moving types, requests and adaptors to proper places
