+++ src/lib/entities/episode/index.ts
@@ -1,1 +1,2 @@
 export { default as EpisodeCardShell } from './ui/episode-card-shell.svelte';
+export { transformEpisodeRequest } from './api/adapters';
