+++ src/lib/entities/podcast/index.ts
@@ -1,2 +1,4 @@
 export { default as PodcastTile } from './ui/podcast-tile.svelte';
 export { default as PodcastDisplay } from './ui/podcast-display.svelte';
+export { transformPodcastRequest } from './api/adapters';
+export { podcastList, podcastGet } from './api/requests';
