+++ src/lib/entities/podcast/ui/specs/podcast-display.spec.ts
@@ -2,7 +2,7 @@ import svelte from 'svelte-inline-compile';
 import { render } from '@testing-library/svelte';
 
 import { useLocalVars } from '$lib/shared/lib/jest-hacks';
-import type { Podcast } from '$lib/shared/api';
+import type { Podcast } from '$lib/shared/types';
 import { PodcastDisplay } from '../..';
 
 const samplePodcast: Podcast = {
