+++ src/lib/features/like-episode/ui/like-button.svelte
@@ -3,7 +3,7 @@
   import IconFavorite from '~icons/ic/sharp-favorite';
   import IconTwoToneFavorite from '~icons/ic/twotone-favorite';
   import { IconButton } from '$lib/shared/ui';
-  import type { Episode } from '$lib/shared/api';
+  import type { Episode } from '$lib/shared/types';
   import { toggleLike, likesStore } from '../model/like';
 
   export let episode: Episode;
