+++ src/lib/pages/library/ui/library-page.svelte
@@ -2,7 +2,7 @@
   import { H1 } from '$lib/shared/ui';
   import { likesStore } from '$lib/features/like-episode';
   import { EpisodeCard } from '$lib/widgets/episode';
-  import type { Episode } from '$lib/shared/api';
+  import type { Episode } from '$lib/shared/types';
 
   $: episodes = Array.from<Episode>($likesStore.values());
 </script>
