+++ src/lib/shared/api/index.ts
@@ -1,3 +1,2 @@
 export { default as supabaseClient } from './supabase';
-export { podcastList, podcastGet } from './podcasts';
-export type { Episode, Podcast } from './types';
+export { notNull } from './not-null';
