+++ src/lib/shared/types/common.d.ts
@@ -1,0 +1,3 @@
+export interface SBFile {
+  name: string;
+}
