+++ src/lib/shared/types/episodes.d.ts
@@ -1,0 +1,14 @@
+export interface SBEpisode {
+  id: number;
+  title: string;
+  duration: number;
+  audio: SBFile;
+}
+
+export interface Episode {
+  id: number;
+  title: string;
+  duration: number;
+  audioUrl: string;
+  favorite?: boolean;
+}
